backend/requirements.txt
flask>=1.0.0
flask_restful
flask-cors
textblob
pocketsphinx 
SpeechRecognition
google-cloud-speech